requirements.txt
# ========================================
<<<<<<< HEAD
# FarmChat - Unified Requirements
=======
# Emergent FarmChat - Unified Requirements
>>>>>>> c7ba531 (Initial push: migrate local codebase to KisanMitr)
# ========================================
# This file consolidates all Python dependencies for the entire project
# Created: October 5, 2025

# ========================================
# BACKEND PYTHON DEPENDENCIES
# ========================================

# Web Framework & API
fastapi==0.110.1
starlette==0.37.2
uvicorn==0.25.0
python-multipart==0.0.20

# Database
motor==3.3.1
pymongo==4.5.0
redis==5.0.1

# Authentication & Security
PyJWT==2.10.1
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.5.0
cryptography==46.0.1

# HTTP Client & Async
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
aiofiles==23.2.1

# Data Processing & Analysis
pandas==2.3.2
numpy==2.3.3

# AI & ML Services
deepgram-sdk==3.7.0

# AWS Services
boto3==1.40.39
botocore==1.40.39
s3transfer==0.14.0
s5cmd==0.2.0

# Caching & Performance
cachetools==5.3.2

# Validation & Data Models
pydantic==2.11.9
pydantic_core==2.33.2
email-validator==2.3.0

# Configuration & Environment
python-dotenv==1.1.1

# Image Processing
Pillow==10.4.0

# Utilities
requests==2.32.5
requests-oauthlib==2.0.0
oauthlib==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
click==8.3.0
watchfiles==1.1.0
dnspython==2.8.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
six==1.17.0
sniffio==1.3.1

# CLI & Terminal
typer==0.19.2
rich==14.1.0
shellingham==1.5.4

# Markdown & Documentation
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2

# Data Handling
jmespath==1.0.1
jq==1.10.0
pytokens==0.1.10

# Cryptography & Security
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1
cffi==2.0.0
pycparser==2.23

# Type Checking & Extensions
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# Development & Testing
pytest==8.4.2
black==25.9.0
flake8==7.3.0
mypy==1.18.2
mypy_extensions==1.1.0
isort==6.0.1
pluggy==1.6.0
iniconfig==2.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
mccabe==0.7.0
anyio==4.11.0

# ========================================
# NOTES
# ========================================
# - Frontend dependencies are in frontend/package.json
# - fs-gate dependencies are in fs-gate/package.json
# - Root package.json only contains dev tooling (shadcn)
#
# Installation:
#   pip install -r requirements.txt
#
# Or for backend only:
#   cd backend && pip install -r requirements.txt
